+++ ChargeLog/Models/DashboardBase.cs
@@ -4,9 +4,9 @@
     {
         public int SessionCount { get; set; }
         public bool DisplayDetails { get; set; } = false;
-        public double KWh { get; set; }
+        public decimal KWh { get; set; }
         public TimeSpan Duration { get; set; }
-        public Double Price { get; set; }
-        public Double Discount { get; set; }
+        public decimal Price { get; set; }
+        public decimal Discount { get; set; }
     }
 }
